[Feature][MRV2] Support MTP
