+++ vllm_ascend/worker/v2/input_batch.py
@@ -100,10 +100,13 @@ class AscendInputBatch(InputBatch):
         # when dummy run for memory profiling,
         # attention metadata isn't needed,
         # we can also set attn_state to AscendAttentionState.DecodeOnly.
-        input_batch.attn_state = AscendAttentionState.DecodeOnly
         # For mla/sfa, update cos/sin. Here is for _dummy_run.
         update_cos_sin(input_batch.positions)
-        return cls(**asdict(input_batch), seq_lens_np=seq_lens_np)
+        return cls(
+            **asdict(input_batch),
+            seq_lens_np=seq_lens_np,
+            attn_state=AscendAttentionState.DecodeOnly,
+        )
 
 
 @triton.jit
