+++ vllm_ascend/worker/v2/spec_decode/autoregressive/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+# SPDX-FileCopyrightText: Copyright contributors to the vLLM-Ascend project
+#
+# Ascend version of the upstream ``autoregressive`` spec_decode module.
+# Hosts the shared ``AscendAutoRegressiveSpeculator`` mixin and the shared
+# aclgraph managers used by both eagle and mtp Ascend speculators.
